+++ packages/components-angular/projects/components-wrapper/src/lib/components-wrapper.component.ts
@@ -177,6 +177,17 @@ export class PPagination {
   }
 }
 
+export declare interface PSelectWrapper extends Components.PSelectWrapper {}
+@ProxyCmp({inputs: ['hideLabel', 'label', 'message', 'state']})
+@Component({ selector: 'p-select-wrapper', changeDetection: ChangeDetectionStrategy.OnPush, template: '<ng-content></ng-content>', inputs: ['hideLabel', 'label', 'message', 'state'] })
+export class PSelectWrapper {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+  }
+}
+
 export declare interface PSpinner extends Components.PSpinner {}
 @ProxyCmp({inputs: ['size', 'theme']})
 @Component({ selector: 'p-spinner', changeDetection: ChangeDetectionStrategy.OnPush, template: '<ng-content></ng-content>', inputs: ['size', 'theme'] })