+++ packages/storefront/storefront.web.config.ts
@@ -47,6 +47,10 @@ export const config: StorefrontConfig = {
       Faq: [
         // @ts-ignore
         () => import(`@/pages/web/faq.md`)
+      ],
+      Troubleshooting: [
+        // @ts-ignore
+        () => import(`@/pages/web/troubleshooting.md`)
       ]
     },
     Basics: {
@@ -177,7 +181,7 @@ export const config: StorefrontConfig = {
         ]
       }
     },
-    Forms: {
+    Form: {
       'Text Field': {
         Design: [
           // @ts-ignore
@@ -190,6 +194,20 @@ export const config: StorefrontConfig = {
         Props: [
           // @ts-ignore
           () => import(`@/../../components-js/src/components/form/text-field-wrapper/text-field-wrapper.props.md`)
+        ]
+      },
+      Select: {
+        Design: [
+          // @ts-ignore
+          () => import(`@/../../components-js/src/components/form/select-wrapper/select-wrapper.design.md`)
+        ],
+        Code: [
+          // @ts-ignore
+          () => import(`@/../../components-js/src/components/form/select-wrapper/select-wrapper.code.md`)
+        ],
+        Props: [
+          // @ts-ignore
+          () => import(`@/../../components-js/src/components/form/select-wrapper/select-wrapper.props.md`)
         ]
       }
     },