+++ packages/components-react/projects/components-wrapper/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `Select Wrapper`
+
 ## [1.0.3] (2020-02-13)
 
 ### Fixed